javadoc + refactoring
