+++ src/main/java/com/exasol/adapter/dynamodb/mapping/tojsonmapping/ToJsonValueMapper.java
@@ -7,13 +7,13 @@ import com.exasol.sql.expression.StringLiteral;
 import com.exasol.sql.expression.ValueExpression;
 
 /**
- * ValueMapper for {@link ToJsonColumnMappingDefinition}
+ * ValueMapper for {@link ToJsonColumnMappingDefinition}.
  */
 @java.lang.SuppressWarnings("squid:S119") // DocumentVisitorType does not fit naming conventions.
 public abstract class ToJsonValueMapper<DocumentVisitorType> extends AbstractValueMapper<DocumentVisitorType> {
 
     /**
-     * Creates an instance of {@link ToJsonColumnMappingDefinition}
+     * Creates an instance of {@link ToJsonValueMapper}
      * 
      * @param column {@link ToJsonColumnMappingDefinition}
      */
